Move `Workflow` and `WorkflowNode` into a sub-package together
