+++ nvtabular/__init__.py
@@ -15,7 +15,7 @@
 #
 import warnings
 
-from . import column_selector, io, workflow, workflow_node  # noqa
+from . import column_selector, io, workflow  # noqa
 from ._version import get_versions
 
 # suppress some warnings with cudf warning about column ordering with dlpack
@@ -24,7 +24,7 @@ warnings.filterwarnings("ignore", module="cudf.io.dlpack")
 warnings.filterwarnings("ignore", module="numba.cuda.envvars")
 
 
-WorkflowNode = workflow_node.WorkflowNode
+WorkflowNode = workflow.WorkflowNode
 ColumnSelector = column_selector.ColumnSelector
 Workflow = workflow.Workflow
 Dataset = io.dataset.Dataset
