+++ tests/conftest.py
@@ -266,7 +266,7 @@ def get_cats(workflow, col, stat_name="categories", cpu=False):
     # figure out the categorify node from the workflow graph
     cats = [
         cg.op
-        for cg in nvtabular.workflow_node.iter_nodes([workflow.workflow_node])
+        for cg in nvtabular.workflow.node.iter_nodes([workflow.workflow_node])
         if isinstance(cg.op, nvtabular.ops.Categorify)
     ]
     if len(cats) != 1:
